test(channel): ChannelService 테스트를 Testcontainers 기반으로 교체
Mockito 단위 테스트를 실물 DB(PostgreSQL Testcontainers)를 사용하는
서비스 통합 테스트로 대체한다.
ChannelEntity columnDefinition 제거 및 Jackson 테스트 의존성 추가.
