+++ infrastructure/build.gradle
@@ -6,6 +6,7 @@ dependencies {
     runtimeOnly    'org.postgresql:postgresql'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    testImplementation 'com.fasterxml.jackson.core:jackson-databind'
     testImplementation 'org.springframework.boot:spring-boot-data-jpa-test'
     testImplementation 'org.springframework.boot:spring-boot-jdbc-test'
     testImplementation 'org.springframework.boot:spring-boot-testcontainers'
